fix crash when ./storage does not initially exist

+++ src/python/network.py
@@ -3,7 +3,8 @@ import multiprocessing
 import time
 import asyncio
 from david.network import Server
-import shutil, os
+import shutil
+import os
 
 from node import Node
 from bnode import BootstrapNode
@@ -15,7 +16,8 @@ class Network():
         self.bnodePort = bnodePort
         self.bnode = BootstrapNode(bnodePort)
         self.nodes = {}  # Map from peer_id of each node to the node itself; makes it easier to call individual node functions
-        shutil.rmtree('./storage')
+        if os.path.exists('./storage'):
+            shutil.rmtree('./storage')
         os.makedirs('./storage')
         worker = multiprocessing.Process(
             target=self.add_job, args=(self.bnode,))
@@ -25,9 +27,11 @@ class Network():
         config['bootstrap_port'] = constants.BOOTSTRAP_PORT
         node = Node(config)
         self.nodes[node.peer_id] = node
-        node_worker = multiprocessing.Process(target=self.add_job, args=(node,))
+        node_worker = multiprocessing.Process(
+            target=self.add_job, args=(node,))
         node_worker.start()
-        dht_worker = multiprocessing.Process(target=self.start_dht_server, args=(config['dht_port'],))
+        dht_worker = multiprocessing.Process(
+            target=self.start_dht_server, args=(config['dht_port'],))
         dht_worker.start()
 
     def start_dht_server(self, port):
@@ -48,6 +52,7 @@ class Network():
     def add_job(self, node):
         t = ThreadedServer(node, port=node.port)
         t.start()
+
 
 async def test(network):
     network.nodes[100].generate_file("hi", 40)
